+++ postcss.config.js
@@ -3,6 +3,6 @@ module.exports = {
         require('autoprefixer'),
         require('cssnano')({ // подключили cssnano
             preset: 'default', // выбрали настройки по умолчанию
-    })
-]
+        })
+    ]
 }
