+++ src/index.html
@@ -10,7 +10,7 @@
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="<%=require('../images/logo.svg')%>" alt="mesto logo" class="logo">
+      <img src="<%=require('../images/logo.svg').default%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -24,26 +24,11 @@
         <button class="button user-info__button">+</button>
       </div>
     </div>
-    <div class="places-list root__section">
-
-      <!-- Пример вёрстки карточки: -->
-<!-- 
-      <div class="place-card">
-        <div class="place-card__image" style="background-image: url(https://pictures.s3.yandex.net/frontend-developer/cards-compressed/kamchatka.jpg)">
-          <button class="place-card__delete-icon"></button>
-        </div>
-        <div class="place-card__description">
-          <h3 class="place-card__name">Камчатка</h3>
-          <button class="place-card__like-icon"></button>
-        </div>
-      </div>
--->
-      <!-- конец примера. -->
-
+    <div class="places-list root__section">    
     </div>
     <div class="popup popup__card">
       <div class="popup__content">
-        <img src="<%=require('../images/close.svg')%>" alt="" class="popup__close">
+        <img src="<%=require('../images/close.svg').default%>" alt="" class="popup__close">
         <h3 class="popup__title">Новое место</h3>
         <form class="popup__form" name="new">
             <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Название">
@@ -56,7 +41,7 @@
     </div>
     <div class="popup popup__data">
       <div class="popup__content">
-        <img src="<%=require('../images/close.svg')%>" alt="" class="popup__close">
+        <img src="<%=require('../images/close.svg').default%>" alt="" class="popup__close">
         <h3 class="popup__title">Редактировать профиль</h3>
         <form class="popup__form" name="data">
           <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Имя">
@@ -68,10 +53,10 @@
   </div>
   <div class="popup popup__enlarge">
     <div class="popup__content popup__image-content">
-      <img src="<%=require('../images/close.svg')%>" alt="" class="popup__close popup__close_image">
+      <img src="<%=require('../images/close.svg').default%>" alt="" class="popup__close popup__close_image">
       <img class="popup__image">
     </div>
   </div>
-  <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
+  <script type="module" src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
 </body>
 </html>
