+++ src/index.js
@@ -1,168 +1,13 @@
 import "./style.css";
+
+import Api from '../blocks/api/api.js';
+import Popup from '../blocks/popup/__form/popup__form.js';
+import PopupEnlarge from '../blocks/popup/__image/popup__image.js';
+import CardList from '../blocks/places-list/places-list.js';
+
 (function () {
 
   const serverUrl = NODE_ENV === 'development' ? 'http://praktikum.tk/cohort5' : 'https://praktikum.tk/cohort5';
-  //class for cards
-  class Card {
-    constructor(image, name) {
-      this._image = image;
-      this._name = name;
-    }
-
-    like() {
-      this.likeIcon.classList.toggle('place-card__like-icon_liked');
-    }
-
-    remove() {
-      this._card.parentElement.removeChild(this._card);
-    }
-
-    onImageClick(func) {
-      this.cardImage.addEventListener('click', () => func(this._image))
-
-    }
-
-    create() {
-      this._card = document.createElement('div'); //add card
-      this._card.classList.add('place-card');
-
-
-      this.cardImage = document.createElement('div'); //add image
-      this.cardImage.classList.add('place-card__image');
-      this._card.appendChild(this.cardImage);
-      this.cardImage.setAttribute('style', 'background-image: url(' + this._image + ')'); //pics from array
-
-
-      this.deleteIcon = document.createElement('button'); //add trash
-      this.deleteIcon.classList.add('place-card__delete-icon');
-      this.cardImage.appendChild(this.deleteIcon);
-      this.deleteIcon.addEventListener('click', (event) => { event.stopPropagation(); this.remove(); });
-
-      const description = document.createElement('div');
-      description.classList.add('place-card__description');
-      this._card.appendChild(description);
-
-      const cardName = document.createElement('h3'); //add name
-      cardName.classList.add('place-card__name');
-      description.appendChild(cardName);
-      cardName.textContent = this._name; //names from array
-
-      this.likeIcon = document.createElement('button'); //add like
-      this.likeIcon.classList.add('place-card__like-icon');
-      description.appendChild(this.likeIcon);
-      this.likeIcon.addEventListener('click', () => this.like());
-
-      return this._card;
-    }
-  }
-
-  //cards list costructor
-  class CardList {
-    constructor(container, cardsArray, popupEnlarge) {
-      this._container = container;
-      this._cardsArray = cardsArray;
-      this._popupEnlarge = popupEnlarge;
-    }
-
-    render() {
-      for (let i = 0; i < this._cardsArray.length; i++) {
-        this.addCard(this._cardsArray[i].link, this._cardsArray[i].name); //add default cards  
-      }
-    }
-
-    addCard(image, name) {
-      const card = new Card(image, name);
-      this._container.appendChild(card.create());
-      card.onImageClick((image) => this._popupEnlarge.open(image));
-    }
-  }
-
-  //popup for image
-  class PopupEnlarge {
-    constructor(container) {
-      this._container = container;
-      this._popupImage = container.querySelector('.popup__image');
-      this._popupClose = container.querySelector('.popup__close_image')
-      this._popupClose.addEventListener('click', () => this.close());
-    }
-
-    open(image) {
-      this._container.classList.add('popup_is-opened');
-      this._popupImage.setAttribute('src', image);
-    }
-
-    close() {
-      this._container.classList.remove('popup_is-opened');
-    }
-  }
-
-  //form popup
-  class Popup {
-    constructor(container) {
-      this._container = container;
-      //hiding the form
-      this._popupClose = container.querySelector('.popup__close');
-      this._popupClose.addEventListener('click', () => this.close());
-
-    }
-
-    open() {
-      this._container.classList.add('popup_is-opened');
-    }
-
-    close() {
-      this._container.classList.remove('popup_is-opened');
-    }
-  }
-
-  class Api {
-    constructor(options) {
-      this.baseUrl = options.baseUrl;
-      this.headers = options.headers;
-    }
-
-    getInitialCards() {
-      return fetch(`${this.baseUrl}/cards`, {
-        headers: this.headers
-      })
-        .then((res) => {
-          if (res.ok) {
-            return res.json();
-          }
-          return Promise.reject(`Ошибка: ${res.status}`);
-        })
-    }
-
-    getUserData() {
-      return fetch(`${this.baseUrl}/users/me`, {
-        headers: this.headers
-      })
-        .then((res) => {
-          if (res.ok) {
-            return res.json();
-          }
-          return Promise.reject(`Ошибка: ${res.status}`);
-        })
-    }
-
-    setUserData(name, job) {
-      return fetch(`${this.baseUrl}/users/me`, {
-        headers: this.headers,
-        method: 'PATCH',
-        body: JSON.stringify({
-          name: name,
-          about: job
-        })
-
-      })
-        .then((res) => {
-          if (res.ok) {
-            return res.json();
-          }
-          return Promise.reject(`Ошибка: ${res.status}`);
-        })
-    }
-  }
 
   const api = new Api({
     baseUrl: serverUrl,
