+++ webpack.config.js
@@ -12,7 +12,7 @@ module.exports = {
     path: path.resolve(__dirname, 'dist'),
     filename: '[name].[chunkhash].js'
   },
-// указали путь к файлу, в квадратных скобках куда вставлять сгенерированный хеш
+  // указали путь к файлу, в квадратных скобках куда вставлять сгенерированный хеш
   module: {
     rules: [
       {
@@ -22,42 +22,42 @@ module.exports = {
           loader: "babel-loader"
         }
       },
-       {
+      {
         test: /\.(eot|ttf|woff|woff2)$/,
         loader: 'file-loader?name=./vendor/[name].[ext]'
-        },
-        {
-            test: /\.(png|jpg|gif|ico|svg)$/,
-            use: [
-                    'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
-                    {
-                            loader: 'image-webpack-loader',
-                            options: {}
-                    },
-            ]
-           },
-           {
-            test: /\.css$/i,
-            use: [
-                (isDev ? 'style-loader' : MiniCssExtractPlugin.loader),
-                'css-loader', 
-                'postcss-loader'
-            ]
-        }
+      },
+      {
+        test: /\.(png|jpg|gif|ico|svg)$/,
+        use: [
+          'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
+          {
+            loader: 'image-webpack-loader',
+            options: {}
+          },
+        ]
+      },
+      {
+        test: /\.css$/i,
+        use: [
+          (isDev ? 'style-loader' : MiniCssExtractPlugin.loader),
+          'css-loader',
+          'postcss-loader'
+        ]
+      }
     ]
   },
-  plugins: [ 
+  plugins: [
     new MiniCssExtractPlugin({
-        filename: 'style.[contenthash].css'
+      filename: 'style.[contenthash].css'
     }),
     new OptimizeCssAssetsPlugin({
-        assetNameRegExp: /\.css$/g,
-        cssProcessor: require('cssnano'),
-        cssProcessorPluginOptions: {
-                preset: ['default'],
-        },
-        canPrint: true
-   }),
+      assetNameRegExp: /\.css$/g,
+      cssProcessor: require('cssnano'),
+      cssProcessorPluginOptions: {
+        preset: ['default'],
+      },
+      canPrint: true
+    }),
     new HtmlWebpackPlugin({
       inject: false,
       template: './src/index.html',
@@ -65,7 +65,7 @@ module.exports = {
     }),
     new WebpackMd5Hash(),
     new webpack.DefinePlugin({
-        'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-       })
+      'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
+    })
   ]
 };
